// ts/dist/watch.d.ts
import type { Build, BuildResult, Log, Run, Canon, ChangeItem, BuildSpec } from './types';
import { FSWatcher } from 'chokidar';
declare class Watch {
    fsw: FSWatcher | undefined;
    wspec: any;
    last?: BuildResult;
    lastChangeTime: number;
    build: Build | undefined;
    runq: Run[];
    doneq: Run[];
    canons: Canon[];
    canonPaths: Set<string>;
    lastrun: Run | undefined;
    idle: number;
    intervalId: ReturnType<typeof setInterval> | undefined;
    startTime: number;
    running: boolean;
    lastChange: ChangeItem;
    lastTrigger: ChangeItem;
    log: Log;
    name: string;
    mode: {
        mod: boolean;
        add: boolean;
        rem: boolean;
    };
    constructor(bspec: BuildSpec, log: Log);
    ensureFSW(): FSWatcher;
    start(initial?: boolean): void;
    canon(path: string): string;
    handleChange(path: string): void;
    drain(): Promise<void>;
    add(path: string): Promise<void>;
    update(br: BuildResult): Promise<void>;
    run(name: string, watch?: boolean, trigger?: string): Promise<BuildResult>;
    stop(): Promise<void>;
    descDeps(deps: Record<string, Record<string, {
        tar: string;
    }>>): string;
}
export { Watch };